pyproject.toml
[tool.poetry]
name = "auctionproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "3.12.2"
Markdown = "^3.3.4"
django-filter = "^21.1"
Pillow = "^8.3.2"
psycopg2-binary = "^2.9.1"
django-money = "^2.1"
django-debug-toolbar = "^3.2.2"
flake8 = "^4.0.1"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.10.0"
celery = "^5.1.2"
redis = "^3.5.3"
django-celery-beat = "^2.2.1"
django = "^3.2.8"
channels = "^3.0.4"
channels_redis = "^3.3.1"
pytest-django = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
